+++ docs-gen/src/resources/helpers/member-title.ts
@@ -3,9 +3,9 @@ import { DeclarationReflection } from 'typedoc';
 export function memberTitle(this: DeclarationReflection) {
   const md = [];
 
-  if (this.flags) {
-    md.push(this.flags.map((flag) => `\`${flag}\``).join(' '));
-  }
+  // if (this.flags) {
+    // md.push(this.flags.map((flag) => `\`${flag}\``).join(' '));
+  // }
   md.push(this.name);
   return md.join(' ');
 }