+++ packages/cubejs-client-react/src/CubeProvider.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import * as PropTypes from 'prop-types';
 import CubeContext from './CubeContext';
 
 const CubeProvider = ({ cubejsApi, children }) => (
@@ -7,10 +6,5 @@ const CubeProvider = ({ cubejsApi, children }) => (
     {children}
   </CubeContext.Provider>
 );
-
-CubeProvider.propTypes = {
-  cubejsApi: PropTypes.object.isRequired,
-  children: PropTypes.any.isRequired
-};
 
 export default CubeProvider;