+++ packages/cubejs-schema-compiler/compiler/UserError.js
@@ -1,3 +1,8 @@
-class UserError extends Error {}
+class UserError extends Error {
+  constructor(message) {
+    super(message);
+    this.type = 'UserError';
+  }
+}
 
 module.exports = UserError;